requirements.txt
matplotlib==3.8.2
numpy==1.24.1
Pillow==9.3.0
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.1
